Remove unused gems
`actionpack-action_caching` adds `caches_action` method
and it was not used once.

`actionpack-page_caching` adds `caches_page` method
and it was not used once.

`coffee-rails` use `.coffee` view, not a single file exists.

`magic_encoding` add support of comments like `# -*- encoding : utf-8 -*-`
and was not used

`ransack` need for search, and there is no such functionality in Wrata,
aslo no calls for `ransack` found

`rails-observers` was not called `ActiveRecord::Observer` or
`ActionController::Caching::Sweeper`, so remove it.

`sqlite3` was used as db a long time ago, not it is PostgreSQL

`sass-rails` should be configured via `config.saas` and no such configure
was maded, also remove it classes